Move GoogleAnalytics component to organisms
